:recycle: Chore: Refactor runBlock into individual files
