+++ packages/vm/src/actions/applyBlock.ts
@@ -1,0 +1,93 @@
+import { Block } from '@tevm/block'
+import { ConsensusType } from '@tevm/common'
+import { bytesToHex } from '@tevm/utils'
+
+import { InternalError, MisconfiguredClientError } from '@tevm/errors'
+import type { BaseVm } from '../BaseVm.js'
+import type { ApplyBlockResult, RunBlockOpts } from '../utils/index.js'
+import { errorMsg } from './errorMsg.js'
+import { applyTransactions } from './applyTransactions.js'
+import { accumulateParentBeaconBlockRoot } from './accumulateParentBeaconBlockRoot.js'
+import { accumulateParentBlockHash } from './accumulateParentBlockHash.js'
+import { assignBlockRewards } from './assignBlockRewards.js'
+import { assignWithdrawals } from './assignWithdrawals.js'
+
+/**
+ * Validates and applies a block, computing the results of
+ * applying its transactions. This method doesn't modify the
+ * block itself. It computes the block rewards and puts
+ * them on state (but doesn't persist the changes).
+ * @param {Block} block
+ * @param {RunBlockOpts} opts
+ */
+export const applyBlock =
+	(vm: BaseVm) =>
+	async (block: Block, opts: RunBlockOpts): Promise<ApplyBlockResult> => {
+		// Validate block
+		if (opts.skipBlockValidation !== true) {
+			if (block.header.gasLimit >= BigInt('0x8000000000000000')) {
+				const msg = errorMsg('Invalid block with gas limit greater than (2^63 - 1)', vm, block)
+				// todo make InvalidBlockError
+				throw new InternalError(msg)
+			}
+			// TODO: decide what block validation method is appropriate here
+			if (opts.skipHeaderValidation !== true) {
+				if (typeof (<any>vm.blockchain).validateHeader === 'function') {
+					await (<any>vm.blockchain).validateHeader(block.header)
+				} else {
+					throw new InternalError('cannot validate header: blockchain has no `validateHeader` method')
+				}
+			}
+			await block.validateData()
+		}
+		if (vm.common.ethjsCommon.isActivatedEIP(4788)) {
+			await accumulateParentBeaconBlockRoot(vm)(
+				block.header.parentBeaconBlockRoot as Uint8Array,
+				block.header.timestamp,
+			)
+		}
+		if (vm.common.ethjsCommon.isActivatedEIP(2935)) {
+			await accumulateParentBlockHash(vm)(block.header.number, block.header.parentHash)
+		}
+
+		const blockResults = await applyTransactions(vm)(block, opts)
+
+		// Add txResult preimages to the blockResults preimages
+		// Also add the coinbase preimage
+
+		if (opts.reportPreimages === true) {
+			if (vm.evm.stateManager.getAppliedKey === undefined) {
+				throw new MisconfiguredClientError(
+					'applyBlock: evm.stateManager.getAppliedKey can not be undefined if reportPreimages is true',
+				)
+			}
+			blockResults.preimages.set(
+				bytesToHex(vm.evm.stateManager.getAppliedKey(block.header.coinbase.toBytes())),
+				block.header.coinbase.toBytes(),
+			)
+			for (const txResult of blockResults.results) {
+				if (txResult.preimages !== undefined) {
+					for (const [key, preimage] of txResult.preimages) {
+						blockResults.preimages.set(key, preimage)
+					}
+				}
+			}
+		}
+
+		if (vm.common.ethjsCommon.isActivatedEIP(4895)) {
+			if (opts.reportPreimages === true) vm.evm.journal.startReportingPreimages?.()
+			await assignWithdrawals(vm)(block)
+			if (opts.reportPreimages === true && vm.evm.journal.preimages !== undefined) {
+				for (const [key, preimage] of vm.evm.journal.preimages) {
+					blockResults.preimages.set(key, preimage)
+				}
+			}
+			await vm.evm.journal.cleanup()
+		}
+		// Pay ommers and miners
+		if (block.common.ethjsCommon.consensusType() === ConsensusType.ProofOfWork) {
+			await assignBlockRewards(vm)(block)
+		}
+
+		return blockResults as any
+	}
