+++ packages/vm/src/actions/index.ts
@@ -1,12 +1,26 @@
 export * from './BlockBuilder.js'
 export * from './BlockStatus.js'
 export * from './BuildStatus.js'
+export * from './DAOConfig.js'
+export * from './accumulateParentBlockHash.js'
+export * from './accumulateParentBeaconBlockRoot.js'
+export * from './applyBlock.js'
+export * from './applyDAOHardfork.js'
+export * from './applyTransactions.js'
+export * from './assignBlockRewards.js'
+export * from './assignWithdrawals.js'
 export * from './buildBlock.js'
+export * from './calculateMinerReward.js'
+export * from './calculateOmmerReward.js'
 export * from './constants.js'
 export * from './deepCopy.js'
-export * from './errorMessage.js'
 export * from './execHardfork.js'
+export * from './genTxTrie.js'
 export * from './generateTxResult.js'
+export * from './parentBeaconBlockRootAddress.js'
+export * from './rewardAccount.js'
 export * from './runBlock.js'
 export * from './runTx.js'
 export * from './txLogsBloom.js'
+export * from './validateRunTx.js'
+export * from './warmAddresses2929.js'
