+++ packages/vm/src/actions/parentBeaconBlockRootAddress.ts
@@ -1,0 +1,3 @@
+import { EthjsAddress } from '@tevm/utils'
+
+export const parentBeaconBlockRootAddress = EthjsAddress.fromString('0x000F3df6D732807Ef1319fB7B8bB8522d0Beac02')
